refactor(app): split CapTableDashboard into nav-named screens
Lift Holdings, Stock classes, Shareholders, Issue stock, Transfer, and
Transactions out of views/ so each COMPANY left-nav item is a module the
dashboard imports. Wallet receipts and write handlers move into
useWalletReceipt and useCapTableWrites. Product behavior is unchanged.
